src/main/java/com/gutterboys/riichi/calculator/yaku/universal/RinshanKaihou.java
package com.gutterboys.riichi.calculator.yaku.universal;

import org.springframework.stereotype.Component;

import com.gutterboys.riichi.calculator.model.RiichiCalculatorRequest;
import com.gutterboys.riichi.calculator.model.PossibleHand;

@Component
public class RinshanKaihou implements UniversalYaku {

    @Override
    public void execute(RiichiCalculatorRequest request, PossibleHand possibleHand) {
        if (request.isDeadWallDraw() && request.isTsumo() && request.getKanCount() > 0) {
            possibleHand.setHan(possibleHand.getHan() + 1);
            possibleHand.getQualifiedYaku().add("Rinshan Kaihou (Dead Wall Draw)");
        }

    }

}
